+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+  * Shows different icon for category and scorecard in check result
+
 ## [1.2.0] - 2023-11-03
 
 ### Added